Implementing supervised learning
I had started the course offered by Andrew NG, DeepLearning.AI which provides the starting point of learning Machine Learning. This course is named "Supervised Machine Learning: Regression and Classification" and provide insights of basic Supervised Learning Models, Unsupervised Learning Models and much more!

In the first week of this course, i.e. Week 1 (labelled as W1), I started off by learning the meaning, applications and fundamentals behind machine learning and then went to to uncover the basics of Univariate Linear Regression Models. I then started to learn about the Mean Squared-Error cost function (used as a quantification of how well our model is performing) along with a fascinating algorithm that minimizes the cost of our model (aka fitting the data/making better approximations for new data) - Gradient Descent, which uses calculus (differentiation) to better our parameters of the model.

In the second week (W2), I started diving into Multiple Linear Regression which systemically takes in multiple "features" of a data point and then makes predictions using the similarly structured algorithm. It inculcates multiple features which still keeping the computation fast using vectorization. Though this was consequently more time taking - and thus to better the speed of reaching the minimum faster, I applied Feature Scaling using min-max normalization, mean normalization and z-score normalization - which made the contours of the cost function (with respect to its parameters) more symmetrical - and thus easier to reach. At the end of the week, I learnt how to fit higher order polynomials to the data we got.

In the final week (W3), I implemented the amazing concept of Logistic Regression used for classification of the given data into two or more pre-trained groups. I went on to visualize the decision boundary and then using it primarily to visualize the classification. To calculate the accurate values of the parameters, I used a new cost function known as the Logistic Loss function (aka Cross-Entropy function) with Gradient Descent. To end up the week, I got to know more about the problems that might occur with these models - Overfitting and Underfitting - problems that cause the curve to fit the data "too well" or "too loosely". To fight it off, I implemented Regularization - a method that penalizes the parameters on their exact basis of dwindling from the best possible curve.

This was all!